+++ NEWS.md
@@ -1,3 +1,5 @@
+# connectcreds (development version)
+
 # connectcreds 0.1.0
 
 * Initial release. `connectcreds` is is a toolkit for making use of credentials